fix(tabs): scrollToElement on mount of View templates
