+++ src/components/templates/carousel/VerticalView.jsx
@@ -5,9 +5,6 @@ import { withRouter } from 'react-router';
 import loadable from '@loadable/component';
 import cx from 'classnames';
 import { RenderBlocks } from '@plone/volto/components';
-import { withScrollToTarget } from '@eeacms/volto-tabs-block/hocs';
-import { getParentTabFromHash } from '@eeacms/volto-tabs-block/helpers';
-
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import '@eeacms/volto-tabs-block/less/carousel.less';
@@ -16,14 +13,7 @@ const Slider = loadable(() => import('react-slick'));
 
 const View = (props) => {
   const slider = React.useRef(null);
-  const [hashlinkOnMount, setHashlinkOnMount] = React.useState(false);
-  const {
-    metadata = {},
-    data = {},
-    tabsList = [],
-    tabs = {},
-    hashlink = {},
-  } = props;
+  const { metadata = {}, data = {}, tabsList = [], tabs = {} } = props;
   const theme = data.theme || 'light';
   const uiContainer = data.align === 'full' ? 'ui container' : false;
 
@@ -38,35 +28,6 @@ const View = (props) => {
     vertical: true,
     verticalSwiping: true,
   };
-
-  React.useEffect(() => {
-    const urlHash = props.location.hash.substring(1) || '';
-    const parentTabId = getParentTabFromHash(data, urlHash);
-    const id = parentTabId;
-    const index = tabsList.indexOf(id);
-    const currentIndex = slider.current?.innerSlider?.state?.currentSlide;
-    const parentId = data.id || props.id;
-    const parent = document.getElementById(parentId);
-    const headerWrapper = document.querySelector('.header-wrapper');
-    const offsetHeight = headerWrapper?.offsetHeight || 0;
-    if (id !== parentId && index > -1 && parent) {
-      if (currentIndex !== index) {
-        slider.current.slickGoTo(index);
-      }
-      setTimeout(() => {
-        const scrollToElement = document.getElementById(urlHash);
-        //TODO: volto now uses react-router-hash-link which automatically scrolls to offset 0
-        props.scrollToTarget(scrollToElement, offsetHeight);
-      }, 100);
-    } else if (id === parentId && parent) {
-      props.scrollToTarget(parent, offsetHeight);
-    }
-
-    if (!hashlinkOnMount) {
-      setHashlinkOnMount(true);
-    }
-    /* eslint-disable-next-line */
-  }, []);
 
   const panes = tabsList.map((tab, index) => {
     return {
@@ -97,5 +58,4 @@ export default compose(
       hashlink: state.hashlink,
     };
   }),
-  withScrollToTarget,
 )(withRouter(View));
